requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
deep-translator==1.11.4
Django==5.1.5
django-ckeditor==6.7.2
django-js-asset==3.0.1
django-redis==5.4.0
djangorestframework==3.15.2
google-trans-new==1.1.9
idna==3.10
iniconfig==2.0.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pytest==8.3.4
pytest-django==4.9.0
redis==5.2.1
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
